Fix issues with ALLOW_TEARING
