+++ wgpu-hal/src/dx12/adapter.rs
@@ -4,7 +4,7 @@ use crate::{
 };
 use std::{mem, sync::Arc, thread};
 use winapi::{
-    shared::{dxgi, dxgi1_2, dxgi1_5, minwindef, windef, winerror},
+    shared::{dxgi, dxgi1_2, windef, winerror},
     um::{d3d12, d3d12sdklayers, winuser},
 };
 
@@ -426,19 +426,8 @@ impl crate::Adapter<super::Api> for super::Adapter {
         };
 
         let mut present_modes = vec![wgt::PresentMode::Mailbox, wgt::PresentMode::Fifo];
-        #[allow(trivial_casts)]
-        if let Some(factory5) = surface.factory.as_factory5() {
-            let mut allow_tearing: minwindef::BOOL = minwindef::FALSE;
-            let hr = factory5.CheckFeatureSupport(
-                dxgi1_5::DXGI_FEATURE_PRESENT_ALLOW_TEARING,
-                &mut allow_tearing as *mut _ as *mut _,
-                mem::size_of::<minwindef::BOOL>() as _,
-            );
-
-            match hr.into_result() {
-                Err(err) => log::warn!("Unable to check for tearing support: {}", err),
-                Ok(()) => present_modes.push(wgt::PresentMode::Immediate),
-            }
+        if surface.supports_allow_tearing {
+            present_modes.push(wgt::PresentMode::Immediate);
         }
 
         Some(crate::SurfaceCapabilities {
